--- Packages/DependenciesHunter/Editor/AllProjectAssetsReferencesWindow.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ae9d6566c17647b99632a3972f959e0b
-timeCreated: 1626364101
