+++ Packages/DependenciesHunter/Editor/DependenciesHunter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ff5c1f059d9a7a849b9e1c377910b82f
+guid: bf9ff4357ec4dd04f9b8e14e7e58456f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
