--- Packages/DependenciesHunter/Editor/ProjectAssetsAnalysisHelper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 4d7fb1b21ddb41c1b5549c8e6814893c
-timeCreated: 1626364149
