--- Packages/DependenciesHunter/Editor/SelectedAssetsAnalysisHelper.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 3e1337ae756e4db9a56bdb8b5edc48e7
-timeCreated: 1626364178
