--- Packages/DependenciesHunter/Editor/SelectedAssetsReferencesWindow.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cafa5563852948af84de397d48cfa620
-timeCreated: 1626364137
